% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{CompareAcrossSamples}
\alias{CompareAcrossSamples}
\title{Title}
\usage{
CompareAcrossSamples(RomaData, Groups, GSThr = 0.001, Mode = "Wil",
  Type = "Over", TestMode = "Aov+Tuk", TestPV1 = 0.05, TestPV2 = 0.05,
  PlotDiag = FALSE, PlotXGSDiff = FALSE)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Groups}{string vector, a vector of group identifier described by strings}

\item{GSThr}{numeric between 0 and 1, the threshould for significance of the test}

\item{Mode}{string, the mode used to detect over- or under-dispersed gene. Can be Willcox ("Wil") or pseudo pv ("PPV)}

\item{Type}{string, wether to use overdispersed ("Over") or underdispersed ("Under") geneset}

\item{TestMode}{string, the type of statistical methodology to assess sample difference. Currentyl only ANOVA + Tukey is implemented ("Aov+Tuk")}

\item{TestPV1}{numeric between 0 and 1, the threshold PV for Anova}

\item{TestPV2}{numeric between 0 and 1, the threshold PV for Tukey}

\item{PlotDiag}{boolean, should diagnostic plot be displyed?}

\item{PlotXGSDiff}{boolean, should all the significant differences by considered?}
}
\description{
Title
}
