% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ConvertModuleNames}
\alias{ConvertModuleNames}
\title{Conversion between gene naming convenctions and organisms of Modulelist}
\usage{
ConvertModuleNames(ModuleList, SourceOrganism = "hsapiens",
  TargetOrganism = "mmusculus", SourceTypes = "Names",
  TargetTypes = "Names", HomologyLevel = 0, HOST = NULL, PATH = NULL)
}
\arguments{
\item{ModuleList}{The module list to be convert.}

\item{SourceOrganism}{Source organism (e.g., "hsapens" or "mmusculus"). Default is "hsapens".}

\item{TargetOrganism}{Target organism (e.g., "hsapens" or "mmusculus"). Default is "mmusculus".}

\item{SourceTypes}{Type of gene names used as input (currently either "Names" or "Ensembl"). Default is "Names".}

\item{TargetTypes}{Type of gene names to be returned (currently either "Names" or "Ensembl"). Default is "Names".}

\item{HomologyLevel}{minimal level of homology (0 by default)}

\item{HOST}{alterntive host is the default one doe snot work}

\item{PATH}{path of the alterntive host}
}
\description{
Conversion between gene naming convenctions and organisms of Modulelist
}
\examples{



}
