% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{DashboardB}
\alias{DashboardB}
\title{Genesets dashboard}
\usage{
DashboardB(RomaData, Groups, Interactive = FALSE)
}
\arguments{
\item{RomaData}{}

\item{Groups}{}

\item{ExpMat}{}
}
\description{
Genesets dashboard
}
\examples{

}
