% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{DetectOutliers}
\alias{DetectOutliers}
\title{Detect outliers}
\usage{
DetectOutliers(GeneOutDetection, GeneOutThr, ModulePCACenter, CompatibleGenes,
  ExpressionData, PlotData = FALSE, ModuleName = "", PrintInfo = TRUE)
}
\arguments{
\item{GeneOutDetection}{}

\item{GeneOutThr}{}

\item{ModulePCACenter}{}

\item{ExpressionData}{}

\item{PlotData}{}

\item{ModuleName}{}

\item{Genes}{}
}
\description{
Detect outliers
}
