% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ExploreGeneProperties}
\alias{ExploreGeneProperties}
\title{Title}
\usage{
ExploreGeneProperties(RomaData, GeneName, ExpressionMatrix, Selected = NULL,
  GroupInfo = NULL)
}
\arguments{
\item{RomaData}{}

\item{GeneName}{}

\item{ExpressionMatrix}{}

\item{Selected}{}

\item{GroupInfo}{}
}
\description{
Title
}
\examples{

}
