% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcsign.R
\name{FixPCSign}
\alias{FixPCSign}
\title{Correct the sign of the principal component}
\usage{
FixPCSign(GeneScore, SampleScore, Wei = NULL, Mode = "none", DefWei = 1,
  Thr = NULL, Grouping = NULL, ExpMat = NULL, CorMethod = "pearson")
}
\arguments{
\item{GeneScore}{}

\item{SampleScore}{}

\item{Wei}{vector, numeric optional vector of gene weigths}

\item{Mode}{scalar, character. Mode to correct the sign}

\item{DefWei}{}

\item{Thr}{scalar, numeric quantile threshold}

\item{Grouping}{}

\item{ExpMat}{}

\item{CorMethod}{c("pearson", "kendall", "spearman")}

\item{PCAType}{character string, the type of PCA to perform. It can be "DimensionsAreGenes" or "DimensionsAreSamples"}
}
\description{
Correct the sign of the principal component
}
