% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{GetCorrelations}
\alias{GetCorrelations}
\title{Title}
\usage{
GetCorrelations(RomaData, Selected = NULL, MatData, Methods = "pearson",
  ConfLevel = 0.95, Construct.Net = FALSE, Construct.Net.PVThr = 0.01)
}
\arguments{
\item{RomaData}{}

\item{Selected}{}

\item{MatData}{}

\item{Methods}{}

\item{ConfLevel}{}
}
\description{
Title
}
