% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{GetSingle}
\alias{GetSingle}
\title{Return its input if it is a single value, else return NA}
\usage{
GetSingle(x)
}
\arguments{
\item{x}{input vector}
}
\description{
Return its input if it is a single value, else return NA
}
