% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{GetTopContrib}
\alias{GetTopContrib}
\title{Get the top contributing genes per geneset}
\usage{
GetTopContrib(RomaData, Selected = NULL, nGenes = 10, OrderType = "Abs",
  ResolveMult = "mean")
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The indices of the genesets to plot}

\item{nGenes}{integer scalar. The number of genes to extract per geneset}

\item{OrderType}{string scalar. The mode of selection for the top contributing genes.
The current implementation allow "Abs" (genes with the largest weight in absolute value),
"Pos" (genes with the largest weight), and "Neg" (genes with the largest negative weight)}

\item{ResolveMult}{string scalar. The name of the function to apply to genes appearing in multiple genesets.
By default it "mean".}
}
\description{
Get the top contributing genes per geneset
}
