% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{GlobalCompareAcrossSamplesGenesets}
\alias{GlobalCompareAcrossSamplesGenesets}
\title{Compare ROMA sample scores across samples from different populations, for each selected geneset}
\usage{
GlobalCompareAcrossSamplesGenesets(
  RomaData,
  Groups,
  Selected = NULL,
  TestMode = "Aov",
  PlotDiag = FALSE
)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Groups}{string vector, a vector of group identifiers. Must contain the names of the samples.}

\item{Selected}{Genesets used to perform the analysis}

\item{TestMode}{string, the type of statistical methodology to assess sample difference. Currently only ANOVA ("Aov") is available}

\item{PlotDiag}{boolean, should diagnostic plot be displayed?}
}
\value{

}
\description{
Compare ROMA sample scores across samples from different populations, for each selected geneset
}
