% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{InferBinaryWeight}
\alias{InferBinaryWeight}
\title{Infer weights from expression data}
\usage{
InferBinaryWeight(ExpressionMatrix, ModuleList, FillAllNA = TRUE)
}
\arguments{
\item{ExpressionMatrix}{}

\item{ModuleList}{}

\item{FillAllNA}{}
}
\value{

}
\description{
Infer weights from expression data
}
