% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{InferBinaryWeigth}
\alias{InferBinaryWeigth}
\title{Infer weigths from expression data}
\usage{
InferBinaryWeigth(ExpressionMatrix, ModuleList, FillAllNA = TRUE)
}
\arguments{
\item{ExpressionMatrix}{}

\item{ModuleList}{}

\item{FillAllNA}{}
}
\description{
Infer weigths from expression data
}
