% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.Genesets}
\alias{Plot.Genesets}
\title{Plot genesets}
\usage{
Plot.Genesets(RomaData, Selected = NULL, GenesetMargin = 4,
  SampleMargin = 4, ColorGradient = colorRamps::blue2red(50),
  cluster_cols = FALSE)
}
\arguments{
\item{RomaData}{}

\item{Selected}{}

\item{GenesetMargin}{}

\item{SampleMargin}{}

\item{ColorGradient}{}

\item{cluster_cols}{}
}
\description{
Plot genesets
}
