% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.Genesets.Samples}
\alias{Plot.Genesets.Samples}
\title{Plot genesets information. Creates a heatmap of sample scores for all selected gene sets. If groups and functions
are specified, creates other heatmaps with data aggregated by groups using the functions provided.}
\usage{
Plot.Genesets.Samples(
  RomaData,
  Selected = NULL,
  GenesetMargin = 4,
  SampleMargin = 4,
  ColorGradient = colorRamps::blue2red(50),
  cluster_cols = FALSE,
  GroupInfo = NULL,
  HMTite = "Selected Genesets",
  AggByGroupsFL = list(),
  Normalize = FALSE,
  Transpose = FALSE,
  ZeroColor = NULL
)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{GenesetMargin}{scalat, integer. The number of rows used to draw the geneset names.}

\item{SampleMargin}{scalat, integer. The number of rows used to draw the sample names.}

\item{ColorGradient}{vector, string. The colors used for the heatmap.}

\item{cluster_cols}{boolean, should the samp^le be reordered according to the dendrogram?}

\item{GroupInfo}{vector, character. A vector describing the group association of each sample.}

\item{HMTite}{scalar, string. The title of the heatmap}

\item{AggByGroupsFL}{list, string. A list of function names (as strings) that will be used to aggregate the
geneset weights and produce additional heatmaps.}

\item{Normalize}{boolean, shuold weights be normalized to c(-1, 1) for each geneset}

\item{Transpose}{boolean, should the samples by plotted on the rows instead of the columns?}

\item{ZeroColor}{string, the color to use to mark the points closed to 0 (e.g., "#FFFFFF")}
}
\value{
A list of matrices containing the aggregated data used to produce the heatmaps.
}
\description{
Plot genesets information. Creates a heatmap of sample scores for all selected gene sets. If groups and functions
are specified, creates other heatmaps with data aggregated by groups using the functions provided.
}
