% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.Genesets.vs.Sampled}
\alias{Plot.Genesets.vs.Sampled}
\title{View genesets L1 and median PC1 compared to null distribution.}
\usage{
Plot.Genesets.vs.Sampled(RomaData, Selected = NULL, Plot = "both")
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{Plot}{string, which parameter should be visualized ? Can be "L1", "PC1Median" or "both"}
}
\value{

}
\description{
View genesets L1 and median PC1 compared to null distribution.
}
