% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.Overdispersed}
\alias{Plot.Overdispersed}
\title{Plot overdispersed genesets}
\usage{
Plot.Overdispersed(RomaData, Thr = 1, Mode = "Wil", GenesetMargin = 4,
  SampleMargin = 4)
}
\arguments{
\item{RomaData}{}

\item{Thr}{}

\item{Mode}{}

\item{GenesetMargin}{}

\item{SampleMargin}{}
}
\description{
Plot overdispersed genesets
}
