% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.Underdispersed}
\alias{Plot.Underdispersed}
\title{Plot underdispersed}
\usage{
Plot.Underdispersed(RomaData, Thr = 1, Mode = "Wil", GenesetMargin = 4,
  SampleMargin = 4)
}
\arguments{
\item{RomaData}{}

\item{Thr}{}

\item{Mode}{}

\item{GenesetMargin}{}

\item{SampleMargin}{}
}
\description{
Plot underdispersed
}
