% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotGeneWeight}
\alias{PlotGeneWeight}
\title{Title}
\usage{
PlotGeneWeight(RomaData, PlotGenes = 40, ExpressionMatrix = NULL,
  LogExpression = TRUE, Selected = NULL, PlotWeigthSign = FALSE)
}
\arguments{
\item{RomaData}{}

\item{PlotGenes}{}

\item{ExpressionMatrix}{}

\item{LogExpression}{}

\item{Selected}{}

\item{PlotWeigthSign}{}
}
\description{
Title
}
