% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotOutliers}
\alias{PlotOutliers}
\title{Plot several plots to investigate outliers}
\usage{
PlotOutliers(
  RomaData,
  ExpressionMatrix,
  Selected = NULL,
  Plot_L1 = TRUE,
  Plot_Weights = FALSE,
  Plot_Projection = FALSE
)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{ExpressionMatrix}{matrix, the input data used to compute rRoma}

\item{Selected}{vector, integer. The position of the genesets to study}

\item{Plot_L1}{boolean, to you want to plot the amount of variance explained by PC1 when a specific gene is removed ?}

\item{Plot_Weights}{boolean, do you want to plot the distribution of gene weights ?}

\item{Plot_Projection}{boolean, do you want to plot the outlier genes in the PC1/PC2 space ?}
}
\value{

}
\description{
Plot several plots to investigate outliers
}
