% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotPCProjections}
\alias{PlotPCProjections}
\title{Plot PC projections score across selected samples}
\usage{
PlotPCProjections(RomaData, Selected = NULL, PlotPCProj = "none")
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{PlotPCProj}{vector, string. The plotting modality of projections. It can containing any combination of the following strings: 'Points', 'Density', or 'Bins'.
Any other value will result in projections not being plotted}
}
\description{
Plot PC projections score across selected samples
}
