% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotRecurringGenes}
\alias{PlotRecurringGenes}
\title{Plot the weigth of genes appearing across multiple genesets}
\usage{
PlotRecurringGenes(RomaData, Selected = NULL, GenesByGroup = 20,
  MinMult = 2)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{GenesByGroup}{scalar, integer. The number of genes per plot}

\item{MinMult}{scalar, integer. The minimal multiplicity to plot}
}
\description{
Plot the weigth of genes appearing across multiple genesets
}
