% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotSampleProjections}
\alias{PlotSampleProjections}
\title{Title}
\usage{
PlotSampleProjections(RomaData, PlotSamples = 40, ExpressionMatrix = NULL,
  LogExpression = TRUE, ModuleToPlot = NULL)
}
\arguments{
\item{RomaData}{}
}
\description{
Title
}
