% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{rRoma.R}
\alias{rRoma.R}
\title{Perform ROMA on a datasets}
\usage{
rRoma.R(ExpressionMatrix, centerData = TRUE, ExpFilter = FALSE, ModuleList,
  UseWeigths = FALSE, DefaultWeight = 1, MinGenes = 10, MaxGenes = 1000,
  ApproxSamples = 2, nSamples = 100, OutGeneNumber = 5, Ncomp = 10,
  OutGeneSpace = 5, FixedCenter = TRUE, GeneOutDetection = "PC1IQR",
  GeneOutThr = 5, GeneSelMode = "All", SampleFilter = FALSE,
  MoreInfo = FALSE, PlotData = FALSE, PCADims = 2, PCSignMode = "none",
  PCSignThr = NULL, UseParallel = FALSE, nCores = NULL,
  ClusType = "PSOCK", SamplingGeneWeights = NULL, FillNAMethod = list())
}
\arguments{
\item{ExpressionMatrix}{matrix, a numeric matrix containing the gene expression information. Columns indicate samples and rows indicated genes.}

\item{centerData}{logical, should the gene expression values be centered over the samples?}

\item{ExpFilter}{logical, should the samples be filtered?}

\item{ModuleList}{list, gene module list}

\item{UseWeigths}{logical, should the weigths be used}

\item{DefaultWeight}{integer scalar, the default weigth to us if no weith is specified by the modile file and an algorithm requiring weigths is used}

\item{MinGenes}{integer, the minimum number of genes reported by a module available in the expression matrix to process the module}

\item{MaxGenes}{integer, the maximum number of genes reported by a module available in the expression matrix to process the module}

\item{ApproxSamples}{integer between 0 and 100 the approximation parameter to reuse samples. This is the minimal percentage variation to reuse samples.
For example 5, means that samples re recalculated only if the number of genes in the geneset has increased by at least 5\%.}

\item{nSamples}{integer, the number of randomized gene sampled (per module)}

\item{OutGeneNumber}{scalar, number of median-absolute-deviations away from median required for the total number of genes expressed in a sample to be called an outlier}

\item{Ncomp}{iteger, number of principal components used to filter samples in the gene expression space}

\item{OutGeneSpace}{scalar, number of median-absolute-deviations away from median required for in a sample to be called an outlier in the gene expression space}

\item{FixedCenter}{logical, should PCA with fixed center be used?}

\item{GeneOutDetection}{character scalar, the algorithm used to filter genes in a module. Possible values are
"L1OutVarPerc" (Percentage variation relative to the median variance explained supporgted by a leave one out approach),
"L1OutVarDC" (Dendrogram clustering statistics on variance explained supported by a leave one out approach), and
"L1OutExpOut" (Number of median-absolute-deviations away from median explined variance). the option "L1OutExpOut" requires the scater package to be installed.}

\item{GeneOutThr}{scalar, threshold used by gene filtering algorithm in the modules. It can represent maximum size of filtered cluster ("L1OutVarDC"), 
minimal percentage variation (L1OutVarPerc) or the number of median-absolute-deviations away from median ("L1OutExpOut")}

\item{GeneSelMode}{character scalar, mode used to sample genes: all available genes ("All") or genes not present in the module ("Others")}

\item{SampleFilter}{logical, should outlier detection be applied to sampled to sample data as well?}

\item{MoreInfo}{logical, shuold detailed information on the computation by printed?}

\item{PlotData}{logical, shuold debugging plots by produced ?}

\item{PCADims}{integer, the number of PCA dimensions to compute. Should be >= 2.
Larger values decrease the error in the estimation of the explained variance but increase the computation time.}

\item{PCSignMode}{characrter scalar, the modality to use to determine the direction of the principal components. The following options are currentlhy available:
'none' (The direction is chosen at random), 'PreferActivation' (the direction is chosen in such a way that the sum of the projection is positive),
'UseAllWeigths' (as 'PreferActivation', but the projections ae multiplied by the weigths, missing weith are set to DefaultWeight),
'UseKnownWeigths' (as 'PreferActivation', but the projections ae multiplied by the weigths, missing weigth are set to 0)}

\item{PCSignThr}{numeric scalar, a quantile threshold to limit the projections to use, e.g., if equal to .9
only the 10\% of genes with the largest projection in absolugte value will be considered.}

\item{UseParallel}{boolean, shuold a parallel environment be used? Note that using a parallel environment will increase the memorey usage as a
copy of the gene expression matrix is needed for each core}

\item{nCores}{integer, the number of cores to use if UseParallel is TRUE. Set to NULL for auto-detection}

\item{ClusType}{string, the cluster type to use. The default value ("PSOCK") should be available on most systems, unix-like environments also support the "PSOCK",
which should be faster.}

\item{SamplingGeneWeights}{named vector, numeric. Weigth so use when correcting the sign of the PC for sampled data.}

\item{FillNAMethod}{names list, additional parameters to pass to the mice function}
}
\description{
Perform ROMA on a datasets
}
