% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rRoma}
\alias{rRoma}
\title{Title}
\usage{
rRoma(ExpressionMatrix, ModuleList, SampleInfo = NULL,
  fillMissingValues = 0, centerData = TRUE, doubleCenterData = FALSE,
  minimalNumberOfGenesInModule = 3, maximalNumberOfGenesInModule = 5000,
  numberOfPermutations = 0, typeOfModuleFile = 0)
}
\arguments{
\item{ExpressionMatrix}{}

\item{ModuleList}{}

\item{SampleInfo}{}

\item{fillMissingValues}{}

\item{centerData}{}

\item{doubleCenterData}{}

\item{minimalNumberOfGenesInModule}{}

\item{maximalNumberOfGenesInModule}{}

\item{numberOfPermutations}{}

\item{typeOfModuleFile}{}
}
\description{
Title
}

