% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{SelectFromInternalDB}
\alias{SelectFromInternalDB}
\title{Return a an internal DB using the provided search string}
\usage{
SelectFromInternalDB(SearchString, BDName = "MsigDB", Version = NULL,
  Mode = "ANY")
}
\arguments{
\item{SearchString}{string vector, the keywords to search}

\item{BDName}{string, scalar. The internal DB to use, can be one of the following
\itemize{
\item 'MsigDB': Molecular signature database (http://software.broadinstitute.org/gsea/msigdb)
\item 'ACSN_Global': ACSN global map (https://acsn.curie.fr)
\item 'ACSN_Apoptosis': ACSN Apoptosis map (https://acsn.curie.fr)
\item 'ACSN_CellCycle': ACSN Cell Cycle map (https://acsn.curie.fr)
\item 'ACSN_DNARepair': ACSN DNA repair map (https://acsn.curie.fr)
\item 'ACSN_EMT': ACSN EMT map (https://acsn.curie.fr)
\item 'ACSN_Survival': ACSN Survival map (https://acsn.curie.fr)
}}

\item{Version}{string scalar, the version of the internal DB to use. If NULL, the most recent version availble in the package will be used}

\item{Mode}{string scalar, the search mode for the keywords.
It can be "ALL" (genesets with all the keywords will be returned) or
"ANY" (genesets with at at least one keyword will be returned)}
}
\description{
Return a an internal DB using the provided search string
}
