% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{SelectGeneSets}
\alias{SelectGeneSets}
\title{Select genesets accoding to specific conditions}
\usage{
SelectGeneSets(RomaData, VarThr = 0.001, VarMode = "Wil",
  VarType = "Over", MedThr = NULL, MedMode = "Wil", MedType = "Over")
}
\arguments{
\item{RomaData}{}

\item{GSThr}{}

\item{Mode}{}

\item{Type}{}
}
\description{
Select genesets accoding to specific conditions
}
