% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{SelectOverDispersedGeneSet}
\alias{SelectOverDispersedGeneSet}
\title{Produce a list of genesets with their overdispersion}
\usage{
SelectOverDispersedGeneSet(GeneExpressionMatrix, nSamples, nGenes,
  GeneOutDetection = "none", GeneOutThr = 1, CenterPCA = FALSE,
  PlotData = FALSE, ModuleName = "", PrintInfo = FALSE, PCADims = 2)
}
\arguments{
\item{GeneExpressionMatrix}{}

\item{nSamples}{}

\item{nGenes}{}

\item{GeneOutDetection}{}

\item{GeneOutThr}{}

\item{CenterPCA}{}

\item{PlotData}{}

\item{ModuleName}{}

\item{PrintInfo}{}
}
\description{
Produce a list of genesets with their overdispersion
}
