% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{CompareAcrossSamples}
\alias{CompareAcrossSamples}
\title{Title}
\usage{
CompareAcrossSamples(RomaData, Groups, Selected = NULL,
  TestMode = "Aov+Tuk", TestPV1 = 0.05, TestPV2 = 0.05,
  PlotDiag = FALSE, PlotXGSDiff = FALSE)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Groups}{string vector, a vector of group identifier described by strings}

\item{Selected}{The genesets used to perform the analysis}

\item{TestMode}{string, the type of statistical methodology to assess sample difference. Currentyl only ANOVA + Tukey is implemented ("Aov+Tuk")}

\item{TestPV1}{numeric between 0 and 1, the threshold PV for the initial test (ANOVA)}

\item{TestPV2}{numeric between 0 and 1, the threshold PV for difference test (Tukey)}

\item{PlotDiag}{boolean, should diagnostic plot be displyed?}

\item{PlotXGSDiff}{boolean, should all the significant differences by considered?}
}
\description{
Title
}
