% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{CompareAcrossSamples}
\alias{CompareAcrossSamples}
\title{Title}
\usage{
CompareAcrossSamples(RomaData, Groups, GSThr = 0.001, Mode = "Wil",
  Type = "Over", TestMode = "Aov+Tuk", TestPV1 = 0.05, TestPV2 = 0.05,
  PlotDiag = FALSE)
}
\arguments{
\item{RomaData}{}

\item{Groups}{}

\item{GSThr}{}

\item{Mode}{}

\item{Type}{}

\item{TestMode}{}

\item{TestPV1}{}

\item{TestPV2}{}

\item{PlotDiag}{}
}
\description{
Title
}
