% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{DashboardA}
\alias{DashboardA}
\title{Sample dashboard}
\usage{
DashboardA(RomaData, Groups, ExpMat, Interactive = FALSE)
}
\arguments{
\item{RomaData}{}

\item{Groups}{}

\item{ExpMat}{}
}
\description{
Sample dashboard
}
\examples{

}
