% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ExploreGeneProperties}
\alias{ExploreGeneProperties}
\title{Title}
\usage{
ExploreGeneProperties(RomaData, GeneName, ExpressionMatrix, Selected = NULL,
  GroupInfo = NULL, DO_tSNE = FALSE, initial_dims = 50, perplexity = 30)
}
\arguments{
\item{RomaData}{}

\item{GeneName}{}

\item{ExpressionMatrix}{}

\item{Selected}{}

\item{GroupInfo}{}

\item{DO_tSNE}{}

\item{initial_dims}{}

\item{perplexity}{}
}
\description{
Title
}
\examples{

}
