% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ExploreGeneProperties}
\alias{ExploreGeneProperties}
\title{Plot several plots to investigate a specific gene}
\usage{
ExploreGeneProperties(
  RomaData,
  GeneName,
  ExpressionMatrix,
  Selected = NULL,
  GroupInfo = NULL,
  PlotExpVar = TRUE,
  PlotExpMean = TRUE,
  PlotExpDist = TRUE,
  PlotWeight = FALSE,
  PlotExpSampleScore = FALSE
)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{GeneName}{string, the name of the gene to study}

\item{ExpressionMatrix}{matrix, the input data used to compute rRoma}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{GroupInfo}{vector, character. A vector describing the group association of each sample.}

\item{PlotExpVar}{boolean. Do you want to plot the expression variance of the gene compared to all other genes in the dataset ?}

\item{PlotExpMean}{boolean. Do you want to plot the expression mean of the gene compared to all other genes in the dataset ?}

\item{PlotExpDist}{boolean. Do you want to plot the distribution of the expression of the gene in all samples ? 
If Groups are supplied, also shows potential significant differences between groups for this gene}

\item{PlotWeight}{boolean. Do you want to plot the gene weight for the considered gene sets, compared to all other gene weights ?}

\item{PlotExpSampleScore}{boolean. Do you want to plot gene expression against sample score for the considered gene sets ?}
}
\value{

}
\description{
Plot several plots to investigate a specific gene
}
