% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{FixPCSign}
\alias{FixPCSign}
\title{Correct the sign of the principal component}
\usage{
FixPCSign(PCWeigth, PCProj, Wei = NULL, Mode = "none", DefWei = 1,
  Thr = NULL, Grouping = NULL, ExpMat = NULL, CorMethod = "pearson")
}
\arguments{
\item{PCWeigth}{}

\item{PCProj}{}

\item{Wei}{vector, numeric optional vector of weigths}

\item{Mode}{scalar, character. Mode to correct the sign}

\item{DefWei}{}

\item{Thr}{scalar, numeric quantile threshold}

\item{Grouping}{}

\item{ExpMat}{}

\item{CorMethod}{c("pearson", "kendall", "spearman")}
}
\description{
Correct the sign of the principal component
}
