% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopContributor.R
\name{GetTopContrib}
\alias{GetTopContrib}
\title{Get the top contributing genes per geneset}
\usage{
GetTopContrib(RomaData, Selected = NULL, nGenes = 0.1, Mode = "Wei",
  Plot = FALSE, ExpressionMatrix = NULL,
  ColorGradient = colorRamps::blue2red(50), cluster_cols = FALSE,
  HMTite = "Top contributing genes", OrderType = "Abs", Transpose = FALSE)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The indices of the genesets to plot}

\item{nGenes}{either a value between 0 and 1 or a positive integer larger than 1. If the parameter
is an integer larger than 1, it will be interpreted as the number of genes to return per geneset.
If the parameter is a value between 1 and 0 it will be interpreted as the ratio or genes to return per
geneset (e.g., .1 indicates 10% of the genes of the geneset)).}

\item{Mode}{string, the mode used to determine the top contributing genes. It can be "Wei"
(gene weigths will be used) or "Cor" (pearson correlation between gene expression and module score will be used)}

\item{Plot}{boolean, shoul the summary heatmap be plotted?}

\item{ExpressionMatrix}{numerical matrix, the expression matrix in the same format used by the rRoma.R function}

\item{ColorGradient}{vector, string. The colors used for the heatmap.}

\item{cluster_cols}{boolean, should the genesets be reordered according to the dendrogram?}

\item{HMTite}{scalar, string. The title of the heatmap.}

\item{OrderType}{string scalar. The mode of selection for the top contributing genes.
The current implementation allow "Abs" (genes with the largest weight in absolute value),
"Pos" (genes with the largest weight), and "Neg" (genes with the largest negative weight)}

\item{Transpose}{boolean, should the genes by plotted on the rows instead of the columns?}
}
\description{
Get the top contributing genes per geneset
}
