% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopContributor.R
\name{GetTopContrib}
\alias{GetTopContrib}
\title{Get the top contributing genes per geneset}
\usage{
GetTopContrib(RomaData, Selected = NULL, nGenes = 0.1, Mode = "Wei",
  Plot = FALSE, ExpressionMatrix = NULL,
  ColorGradient = colorRamps::blue2red(50), cluster_cols = FALSE,
  HMTite = "Top contributing genes", OrderType = "Abs", Transpose = FALSE)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The indices of the genesets to plot}

\item{nGenes}{integer scalar. The number of genes to extract per geneset}

\item{ColorGradient}{vector, string. The colors used for the heatmap.}

\item{cluster_cols}{boolean, should the genesets be reordered according to the dendrogram?}

\item{HMTite}{scalar, string. The title of the heatmap.}

\item{OrderType}{string scalar. The mode of selection for the top contributing genes.
The current implementation allow "Abs" (genes with the largest weight in absolute value),
"Pos" (genes with the largest weight), and "Neg" (genes with the largest negative weight)}

\item{Transpose}{boolean, should the genes by plotted on the rows instead of the columns?}
}
\description{
Get the top contributing genes per geneset
}
