% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{GetTopContrib}
\alias{GetTopContrib}
\title{Title}
\usage{
GetTopContrib(RomaData, Thr = 1, Mode = "Wil", Type = "Over",
  nGenes = 10, OrderType = "Abs")
}
\arguments{
\item{RomaData}{}

\item{Thr}{}

\item{Mode}{}

\item{Type}{}

\item{nGenes}{}

\item{OrderType}{}
}
\description{
Title
}
