% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\name{GlobalCompareAcrossSamples}
\alias{GlobalCompareAcrossSamples}
\title{Compare ROMA scores across samples from different populations}
\usage{
GlobalCompareAcrossSamples(
  RomaData,
  Groups,
  Selected = NULL,
  TestMode = "Aov+Tuk",
  TestPV1 = 0.05,
  TestPV2 = 0.05,
  PlotDiag = FALSE,
  PlotXGSDiff = FALSE,
  Plot = "both"
)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Groups}{string vector, a vector of group identifiers. Must contain the names of the samples.}

\item{Selected}{Genesets used to perform the analysis}

\item{TestMode}{string, the type of statistical methodology to assess sample difference. Currently only ANOVA + Tukey is implemented ("Aov+Tuk")}

\item{TestPV1}{numeric between 0 and 1, the threshold PV for the initial test (ANOVA)}

\item{TestPV2}{numeric between 0 and 1, the threshold PV for difference test (Tukey)}

\item{PlotDiag}{boolean, should diagnostic plot be displayed?}

\item{PlotXGSDiff}{boolean, should all significant differences by considered?}

\item{Plot}{string, whether difference should be plotted by group or by sample. Possible values are "group", "sample", "both"}
}
\value{

}
\description{
Compare ROMA scores across samples from different populations
}
