% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.Genesets}
\alias{Plot.Genesets}
\title{Plot genesets information}
\usage{
Plot.Genesets(RomaData, Selected = NULL, GenesetMargin = 4,
  SampleMargin = 4, ColorGradient = colorRamps::blue2red(50),
  cluster_cols = FALSE, GroupInfo = NULL)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{GenesetMargin}{scalat, integer. The number of rows used to draw the geneset names.}

\item{SampleMargin}{scalat, integer. The number of rows used to draw the sample names.}

\item{ColorGradient}{vector, string. The colors used for the heatmap.}

\item{cluster_cols}{boolean, should the samp^le be reordered according to the dendrogram?}

\item{GroupInfo}{vector, character. A vector describing the group association of each sample}
}
\description{
Plot genesets information
}
