% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotGeneWeight}
\alias{PlotGeneWeight}
\title{Title}
\usage{
PlotGeneWeight(RomaData, PlotGenes = 40, ExpressionMatrix = NULL,
  LogExpression = TRUE, ModuleToPlot = NULL)
}
\arguments{
\item{RomaData}{}
}
\description{
Title
}
