% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotGeneWeight}
\alias{PlotGeneWeight}
\title{Plot gene weight across selected samples}
\usage{
PlotGeneWeight(
  RomaData,
  PlotGenes = 40,
  ExpressionMatrix = NULL,
  LogExpression = TRUE,
  Selected = NULL,
  PlotWeightSign = FALSE
)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{PlotGenes}{scalar, numeric. The number of genes to plot}

\item{ExpressionMatrix}{matrix, numeric. The expression matrix used to produce gene expression boxplot. If NULL (default), no gene expression information is reported}

\item{LogExpression}{boolean, should gene expression be logtransformed?}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{PlotWeightSign}{boolean, should the sign of the genes weight be used to color the plots?}
}
\value{

}
\description{
Plot gene weight across selected samples
}
