% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{PlotOnACSN}
\alias{PlotOnACSN}
\title{Title}
\usage{
PlotOnACSN(RomaData, SampleName, AggSampFun = "mean", MapURL = "",
  Selected = NULL, FilterByWei = NULL, AggGeneFun = "mean",
  DispMode = "Module", DataName = "Sample")
}
\arguments{
\item{RomaData}{}

\item{SampleName}{}

\item{AggSampFun}{}

\item{Selected}{}

\item{FilterByWei}{}

\item{AggGeneFun}{}

\item{DispMode}{}

\item{DataName}{}
}
\description{
Title
}
