% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotSampleProjections}
\alias{PlotSampleProjections}
\title{Title}
\usage{
PlotSampleProjections(RomaData, PlotSamples = 40, ExpressionMatrix = NULL,
  LogExpression = TRUE, Selected = NULL)
}
\arguments{
\item{RomaData}{}

\item{PlotSamples}{}

\item{ExpressionMatrix}{}

\item{LogExpression}{}

\item{Selected}{}
}
\description{
Title
}
