% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotSampleProjections}
\alias{PlotSampleProjections}
\title{Plot gene weigth across selected samples}
\usage{
PlotSampleProjections(RomaData, PlotSamples = 40, ExpressionMatrix = NULL,
  LogExpression = TRUE, Selected = NULL, PlotPCProj = "none")
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{PlotSamples}{scalar, numeric. The number of samples to plot}

\item{ExpressionMatrix}{matrix, numeric. The expression matrix used to produce gene expression boxplot. If NULL (default), no gene expression information is reported}

\item{LogExpression}{boolean, should gene expression be logtransformed?}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{PlotPCProj}{vector, string. The plotting modality of projections. It can containing any combination of the following strings: 'Points', 'Density', or 'Bins'.
Any other value will result in projections not being plotted}
}
\description{
Plot gene weigth across selected samples
}
