% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotSampleProjections}
\alias{PlotSampleProjections}
\title{Plot sample score across selected samples}
\usage{
PlotSampleProjections(
  RomaData,
  PlotSamples = 40,
  ExpressionMatrix = NULL,
  LogExpression = TRUE,
  Selected = NULL,
  FullExpDist = FALSE
)
}
\arguments{
\item{RomaData}{list, the analysis returned by rRoma}

\item{PlotSamples}{scalar, numeric. The number of samples to plot}

\item{ExpressionMatrix}{matrix, numeric. The expression matrix used to produce gene expression boxplot. If NULL (default), no gene expression information is reported}

\item{LogExpression}{boolean, should gene expression be logtransformed?}

\item{Selected}{vector, integer. The position of the genesets to plot}

\item{FullExpDist}{boolean, should the all the genes be used when showing gene expression
distribution (TRUE) or only the genes of the geneset under consideration (FALSE) ?}
}
\value{

}
\description{
Plot sample score across selected samples
}
