% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{rRoma.R}
\alias{rRoma.R}
\title{Perform ROMA on a datasets}
\usage{
rRoma.R(ExpressionMatrix, centerData = TRUE, ExpFilter = FALSE, ModuleList,
  UseWeigths = FALSE, DefaultWeight = 1, MinGenes = 10, MaxGenes = 1000,
  ApproxSamples = 2, nSamples = 100, OutGeneNumber = 5, Ncomp = 10,
  OutGeneSpace = 5, FixedCenter = TRUE, GeneOutDetection = "PC1IQR",
  GeneOutThr = 5, GeneSelMode = "All", SampleFilter = FALSE,
  MoreInfo = FALSE, PlotData = FALSE, PCADims = 2, PC1SignMode = "none",
  PC1SignThr = NULL)
}
\arguments{
\item{ExpressionMatrix}{matrix, a numeric matrix containing the gene expression information. Columns indicate samples and rows indicated genes.}

\item{centerData}{logical, should the gene expression values be centered over the samples?}

\item{ExpFilter}{logical, should the samples be filtered?}

\item{ModuleList}{list, gene module list}

\item{UseWeigths}{logical, should the weigths be used}

\item{DefaultWeight}{integer scalar, the default weigth to us if no weith is specified by the modile file and an algorithm requiring weigths is used}

\item{MinGenes}{integer, the minimum number of genes reported by a module available in the expression matrix to process the module}

\item{MaxGenes}{integer, the maximum number of genes reported by a module available in the expression matrix to process the module}

\item{ApproxSamples}{integer, the approximation parameter to reuse samples}

\item{nSamples}{integer, the number of randomized gene sampled (per module)}

\item{OutGeneNumber}{scalar, number of median-absolute-deviations away from median required for the total number of genes expressed in a sample to be called an outlier}

\item{Ncomp}{iteger, number of principal components used to filter samples in the gene expression space}

\item{OutGeneSpace}{scalar, number of median-absolute-deviations away from median required for in a sample to be called an outlier in the gene expression space}

\item{FixedCenter}{logical, should PCA with fixed center be used?}

\item{GeneOutDetection}{character scalar, the algorithm used to filter genes in a module. Possible values are "PC1IQR" (Projection on PC1, order statistics),
"PC1DC" (Projection on PC1, dendrogram clustering), "L1OutVarPerc" (Percentage variation relative to the median variance explained supporgted by a leave one out approach), and
"L1OutVarDC" (Dendrogram clustering statistics on variance explained supported by a leave one out approach)}

\item{GeneOutThr}{scalar, threshold used by gene filtering algorithm in the modules. It can represent maximum size of filtered cluster ("PC1DC" and "L1OutVarDC"),
minimal percentage variation (L1OutVarPerc), or minimal median-absolute-deviations.}

\item{GeneSelMode}{character scalar, mode used to sample genes: all available genes ("All") or genes not present in the module ("Others")}

\item{SampleFilter}{logical, should outlier detection be applied to sampled to sample data as well?}

\item{MoreInfo}{logical, shuold detailed information on the computation by printed?}

\item{PlotData}{logical, shuold debugging plots by produced ?}

\item{PCADims}{integer, the number of PCA dimensions to compute. Should be >= 2.
Larger values decrease the error in the estimation of the explained variance but increase the computation time.}

\item{PC1SignMode}{characrter scalar, the modality to use to determine the direction of the 1st principal component. The following options are currentlhy available:
'none' (The direction is chosen at random), 'PreferActivation' (the direction is chosen in such a way that the sum of the projection is positive),
'UseAllWeigths' (as 'PreferActivation', but the projections ae multiplied by the weigths, missing weith are set to DefaultWeight),
'UseKnownWeigths' (as 'PreferActivation', but the projections ae multiplied by the weigths, missing weigth are set to 0)}

\item{PC1SignThr}{numeric scalar, a quantile threshold to limit the projections to use, e.g., if equal to .9
only the 10\% of genes with the largest projection in absolugte value will be considered.}
}
\description{
Perform ROMA on a datasets
}
