% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{SelectFromMSIGdb}
\alias{SelectFromMSIGdb}
\title{Return a subset of Msigdb containing the provided search string}
\usage{
SelectFromMSIGdb(SearchString)
}
\arguments{
\item{SearchString}{string scalar, the keyword to search}
}
\description{
Return a subset of Msigdb containing the provided search string
}
